Refactor catserver updater tests
